requirements/base.txt
# Django
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5

# Database
psycopg2-binary==2.9.9
redis==5.0.1

# WeChat SDK
wechatpy==1.8.18

# Payment
WeChatPy==1.8.18

# Utils
python-dotenv==1.0.0
Pillow==10.1.0
openpyxl==3.1.2
cryptography==41.0.7

# API Documentation
drf-yasg==1.21.7

# Task Queue
celery==5.3.4
django-celery-beat==2.5.0

# Security
djangorestframework-simplejwt==5.3.0

# Deployment
gunicorn==21.2.0
whitenoise==6.6.0
